+++ SocketCommunication/CommandReceiverMissionControl.py
@@ -1,11 +1,12 @@
 from Domain.Communication.CommunicationAbstraction import CommandReceiver
-from Domain.MissionRover.instruction import Instruction
+from Domain.MissionRover.command import Command
 
+# Classe qui reçoit les données envoyées par MissionControl
 class CommandReceiverMissionControl(CommandReceiver):
     def receive_command(self, protocol):
         data = protocol.receive_data_with_length(protocol)
 
         # Décodage des données
-        instructions_received = data.decode('utf-8')
-        instructions = Instruction(instructions_received, True)
-        return instructions
+        commands_received = data.decode('utf-8')
+        commands = Command(commands_received, True)
+        return commands
