+++ SocketCommunication/CommandSenderRover.py
@@ -1,7 +1,7 @@
-# Envoie des données à Rover
 from Domain.Communication.CommunicationAbstraction import CommandSender
 
+# Classe qui envoie des données à Rover
 class CommandSenderRover(CommandSender):
-    def send_command(self, protocol, instructions):
-        data_to_send = protocol.encode(instructions)
+    def send_command(self, protocol, commands):
+        data_to_send = protocol.encode(commands)
         protocol.client_socket.sendall(data_to_send)
